Rework run config generation.
